Add files via upload
🎉 SQL Achievement Unlocked: 5 Gold Stars on HackerRank! 🌟🌟🌟🌟🌟

I am thrilled to announce that I've achieved a significant milestone in my SQL journey on HackerRank - 5 Gold Stars! 🚀✨

This accomplishment wouldn't have been possible without the countless hours of learning, problem-solving, and determination. I faced challenges, learned from mistakes, and grew as a SQL enthusiast.

To those still on the SQL learning journey, keep pushing forward! Remember, each challenge is an opportunity to learn and improve. Celebrate the small victories, and don't be afraid to tackle the tough problems.

Thank you, HackerRank, for providing a platform that continually challenges and inspires us to excel in our skills. Here's to more learning, more challenges, and more gold stars in the future!

#SQLMaster #HackerRank #AchievementUnlocked #SQLJourney 🎉🌟🔍📊👨‍💻